Revert "fixes the problem of tag <input type='file'> without attribute accept=' not allowing user to choose to take a picture or video with the camera. Before it only allowed file upload without explicitly declaring the mime types in accept attribute now it will allow the option of taking picture or video as most browsers do.  Can be tested at: https://andreipfeiffer.github.io/react-native-webview-android-file-upload/index.html"
This reverts commit 9d125b807382d102351b3d5db42f1c378e4f234a.

I am doing some other changes that will allow for uploads of files that
are not images or videos when the accepts attribute is not set.
This means it fixes the problem of */* being entered into the array and so
this commit can be pulled.. will commit with the full changes.