fix(Android): Update `activeUrl` before `goBack` so that `onLoadProgress` could reflect the complete progress when user navigates to the previous page
`shouldOverrideUrlLoading` is not called for the previous page when user presses the back button on phone, therefore `activeUrl` is not pointing to the correct url. This causes `onLoadProgress` not going through a full cycle, the last call would only have a progress of about 0.1. Set the correct `activeUrl` before executing the `goBack` command solves this problem.