Improve onLoadProgress consistency
Currently `onLoadProgress` callback is very inconsistent, depend on the way the url is loaded, the callback might not be called with progress from a low number all the way to `1`.
Many of the inconsistencies come from react-native-community/react-native-webview#643 and the order of the webview callbacks(onPageStarted, onPageFinished, shouldOverrideUrlLoading, onProgressChanged) are called.

Details of the original issue(react-native-community/react-native-webview#642) react-native-community/react-native-webview#643 was trying to solve:
We always return true to `shouldOverrideUrlLoading` so that we can asynchronously decide whether to load the url or not. But the first overridden synchronous load also creates additional call to `onLoadProgress`. We want to filter out those calls.

This commit uses a different way to tackle with the original issue(react-native-community/react-native-webview#642) and get rid of the inconsistencies:
Use a variable `isOverriddenLoad` to mark the first load(which is going to be overridden), and don't call `onLoadProgress` on the load. Only call `onLoadProgress` if we later in Javascript side decide to load this url.