feat(apple): userAgent can be changed at runtime
Make it so that if you change the userAgent prop, this change is propagated to
the native webview on iOS/macOS. Android already has this functionality.

Sometimes it is useful for WebView apps to be able to change their user agent
dynamically to communicate custom things to the web app they are hosting.
Since we can't add custom headers to every single request made by the webview,
the user agent can be used to send things like feature flags to the web app, so
it can have the appropriate behavior depending on what state the mobile app is
in.