docs: add more details on onContentProcessDidTerminate
We were logging every call to `onContentProcessDidTerminate` as an unexpected error on our app. 

Turns out it was completely expected to happen at some point. 
I added more details from [the Apple docs](https://developer.apple.com/documentation/webkit/wknavigationdelegate/1455639-webviewwebcontentprocessdidtermi) to make users aware that WKWebView live in completely different processes than the app. As such they are not included in the total RAM of the app, and can be terminated independently of the app to liberate some new apps.